+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added direct message to a user functionality
 - Added endpoint to list all tags
+- Added endpoint to create categories
 
 ## v0.8 (2018-10-18)